+++ demos/demo/lib/screens/home_screen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import 'lx_computed_demo.dart';
 import 'lx_data_builder_demo.dart';
 import 'lx_multi_builder_demo.dart';
 import 'lx_operators_demo.dart';
@@ -65,6 +66,14 @@ class HomeScreen extends StatelessWidget {
               onPressed: () {
                 Navigator.of(context).push(MaterialPageRoute(
                     builder: (_) => LxOperatorsDemoView()));
+              },
+            ),
+            // Computed Values
+            ElevatedButton(
+              child: Text('Computed Values Demo'),
+              onPressed: () {
+                Navigator.of(context).push(MaterialPageRoute(
+                    builder: (_) => LxComputedDemoView()));
               },
             ),
           ],